+++ src/main/java/com/grupparbete/repositories/UserSqlRepository.java
@@ -3,6 +3,9 @@ package com.grupparbete.repositories;
 import com.grupparbete.entities.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserSqlRepository extends JpaRepository<User, Integer> {
+import java.util.List;
+
+public interface UserSqlRepository extends JpaRepository<User, Long> {
+   // List<User>getById(Long id);
 
 }
